+++ Genesis/src/utopia/genesis/shape/path/Path.scala
@@ -3,12 +3,36 @@ package utopia.genesis.shape.path
 import utopia.genesis.animation.Animation
 import utopia.genesis.util.DistanceLike
 
+object Path
+{
+	// TYPES    -----------------
+	
+	/**
+	  * A common type of path that knows its length
+	  */
+	type PathWithDistance[+X] = Path[X] with DistanceLike
+	
+	
+	// EXTENSIONS   -------------
+	
+	implicit class CombinablePath[A](val p: PathWithDistance[A]) extends AnyVal
+	{
+		/**
+		  * Continues this path with another
+		  * @param another Another path
+		  * @tparam B The type of resulting path
+		  * @return A path that starts with this path and continues with the another
+		  */
+		def +[B >: A](another: PathWithDistance[B]) = CompoundPath(Vector(p, another))
+	}
+}
+
 /**
   * Paths form a sequence of points. They have a specified start and end point
   * @author Mikko Hilpinen
   * @since 19.6.2019, v2.1+
   */
-trait Path[+P] extends Animation[P] with DistanceLike
+trait Path[+P] extends Animation[P]
 {
 	// ABSTRACT	----------------
 	
@@ -20,17 +44,6 @@ trait Path[+P] extends Animation[P] with DistanceLike
 	  * @return The end point of this path
 	  */
 	def end: P
-	
-	
-	// OPERATORS	-------------
-	
-	/**
-	  * Continues this path with another
-	  * @param another Another path
-	  * @tparam B The type of resulting path
-	  * @return A path that starts with this path and continues with the another
-	  */
-	def +[B >: P](another: Path[B]) = CompoundPath(Vector(this, another))
 	
 	
 	// OTHER	-----------------